refactor: it separates the test contexts.
